commit supervised_llc functional and successful training code

+++ utils/dataset.py
@@ -4,8 +4,8 @@ import os
 
 class SupervisedLLCDataset(torch.utils.data.Dataset):
     def __init__(self, datapath):
-        X = torch.from_numpy(np.load(os.path.join(datapath, 'X.npy')))
-        Y = torch.from_numpy(np.load(os.path.join(datapath, 'Y.npy')))
+        X = torch.from_numpy(np.load(os.path.join(datapath, 'X.npy')).astype('float32'))
+        Y = torch.from_numpy(np.load(os.path.join(datapath, 'Y.npy')).astype('float32'))
         self.data = list(zip(X, Y))
 
     def __len__(self):
